// types.ts
export interface DepthLevel {
    price: number;
    orders: number;
    quantity: number;
}

export interface MarketDepth {
    bids: DepthLevel[];
    asks: DepthLevel[];
}

export interface OverviewAsset {
    symbol: string;
    name: string;
    ltp: number;
    change: number;
    changePercent: number;
    history: { name: string; value: number }[];
}

export enum InstrumentType {
    EQUITY = 'EQUITY',
    INDEX = 'INDEX',
    FUTURE = 'FUTURE',
    OPTION = 'OPTION'
}

export interface Stock {
    symbol: string;
    name: string;
    exchange: string;
    instrumentType?: InstrumentType; // New field
    underlying?: string; // For F&O (e.g., NIFTY 50)
    expiryDate?: string; // For F&O
    strikePrice?: number; // For Options
    optionType?: 'CE' | 'PE'; // For Options
    ltp: number; // Last Traded Price
    open: number;
    high: number;
    low: number;
    prevClose: number;
    change: number;
    changePercent: number;
    event?: string;
    bid?: number;
    ask?: number;
    marketDepth?: MarketDepth;
    volume?: number;
    avgTradePrice?: number;
    lastTradedQuantity?: number;
    lastTradedAt?: string;
    lowerCircuit?: number;
    upperCircuit?: number;
}

// Incoming Data from Broadcasting WebSocket
export interface TickData {
    symbol: string;
    ltp: number;
    change?: number;
    percentChange?: number;
    high?: number;
    low?: number;
    open?: number;
    volume?: number;
    timestamp?: number;
}